% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPcluster.R
\name{MAPcluster}
\alias{MAPcluster}
\title{Extract the Maximum A Posteriori cluster estimator from the posterior sample.}
\usage{
MAPcluster(
  voteCube = voteCube,
  nclust_store = dynResult$nclust_store_data,
  clusterSample = dynResult$c_store,
  dataStructure = dataStructure
)
}
\description{
Extract the Maximum A Posteriori cluster estimator from the posterior sample.
}
