\name{MCMCddpmbb-package}
\alias{MCMCddpmbb-package}
\alias{MCMCddpmbb}
\docType{package}
\title{
\packageTitle{MCMCddpmbb}
}
\description{
\packageDescription{MCMCddpmbb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MCMCddpmbb}
\packageIndices{MCMCddpmbb}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MCMCddpmbb}

Maintainer: \packageMaintainer{MCMCddpmbb}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
