% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCMCddpmbb}
\alias{MCMCddpmbb}
\title{Dynamic Dirichlet Process Mixture Beta-Binomial Model to Indentify Voting Coalitions in Roll Call Vote Data of Multiple Periods.}
\usage{
MCMCddpmbb(
  X,
  dataStructure,
  lambda1_value,
  lambda0_value,
  alpha_start,
  tau,
  W = 1,
  m0 = 0,
  H0 = 1,
  r0 = 1,
  s0 = 1,
  V_start = 1,
  sams_iter = 3L,
  mcmc = 1000L,
  burnin = 0L,
  thin = 1L,
  verbose = 100L,
  seed = 1234L,
  fix_W = 1L,
  r1 = 0.1,
  s1 = 1
)
}
\description{
Dynamic Dirichlet Process Mixture Beta-Binomial Model to Indentify Voting Coalitions in Roll Call Vote Data of Multiple Periods.
}
