% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modecluster.R
\name{Modecluster}
\alias{Modecluster}
\title{show the mode of the numbers of clusters across iterations.}
\usage{
Modecluster(nclust_store = dynResult$nclust_store_data)
}
\description{
show the mode of the numbers of clusters across iterations.
}
