% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_probability_same_cluster_heatmap.R
\name{pair_probability_same_cluster_heatmap}
\alias{pair_probability_same_cluster_heatmap}
\title{show the posterior probability for a unique voter pair to be in the same cluster.}
\usage{
pair_probability_same_cluster_heatmap(
  voteCube_t,
  clusters_mat,
  statesName_t_iso,
  n_voter,
  n_item,
  font_size = 5
)
}
\arguments{
\item{voteCube_t}{the vote matrix in a period t}
}
\description{
show the posterior probability for a unique voter pair to be in the same cluster.
}
